requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
sql-migrate==0.1.0
SQLAlchemy==1.3.20
